[FE] rename: DatePicker components로 이동
- 기존 GroupCreatePage에 있었던 DatePicker 컴포넌트를 다른곳에서 재사용되기 때문에 components 디렉토리로 이동
