+++ client/src/components/index.ts
@@ -10,3 +10,4 @@ export { default as SearchBar } from './SearchBar';
 export { default as GroupCard } from './GroupCard';
 export { default as Pagination } from './Pagination';
 export { default as MentorCard } from './MentorCard';
+export { default as DatePicker} from './DatePicker';
