+++ client/src/pages/GroupCreatePage/DateInput/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { useAppDispatch, useAppSelector } from '@hooks';
 import { groupCreateDataState, setGroupData } from '@store/group/makerSlice';
-import AntDatePicker from './AntDatePicker';
+import { DatePicker } from '@components';
 
 function DateInput(): JSX.Element {
   const { startAt, endAt } = useAppSelector(groupCreateDataState);
@@ -14,7 +14,7 @@ function DateInput(): JSX.Element {
   return (
     <>
       <Title>시작일, 종료일을 선택해주세요.</Title>
-      <AntDatePicker startAt={startAt} endAt={endAt} setDate={setDate} />
+      <DatePicker startAt={startAt} endAt={endAt} setDate={setDate} />
     </>
   );
 }
